Split ObjectAssertionSpecs in multiple files similar to StringAssertionSpecs in #1903
